Do not test for clean namespaces
Our code does an odd thing where models generate attributes in consuming
classes. To do this they call $class->can('has') to get a subref they can
call. If we use namespace::autoclean the ->can call returns undef.

We could fix this by passing the subref in, but we'd need to fix the GeoIP2
package to accept this at the same time.